fix(module imports): moved functionality that belonged to just the addon, over to the addon. fix #259
